feat(client): use protocol connector
