+++ client/src/connect/mod.rs
@@ -1,11 +1,10 @@
+use protocol::DomainOrSocketAddr;
 use tokio::io::{AsyncRead, AsyncWrite};
-
-use crate::addr::domain_or_socket_addr::DomainOrSocketAddr;
 
 pub mod hyper;
 
-#[trait_variant::make(Connect: Send)]
-pub trait LocalConnect {
+#[trait_variant::make(Send)]
+pub trait Connect {
     type Read: AsyncRead + Unpin + Send + 'static;
     type Write: AsyncWrite + Unpin + Send + 'static;
